pyproject.toml
[tool.poetry.scripts]
flow = 'gregtech:main'

[tool.poetry]
name = "gregtech-flow"
packages = [
    {include = "gregtech", from="./src"}
]
version = "0.4.5"
description = "Factory Optimization Flowcharts for Gregtech: New Horizons"
authors = ["velolib <vlocitize@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://velolib.github.io/gregtech-flow"
repository = "https://github.com/velolib/gregtech-flow"
documentation = "https://velolib.github.io/gregtech-flow"
keywords = ["gregtech", "minecraft", "graphviz", "graphs"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Intended Audience :: End Users/Desktop"
]

[tool.poetry.dependencies]
python = "^3.11"
graphviz = "^0.20.1"
rich = "^13.0.0"
sympy = "^1.11.1"
typer = "^0.7.0"
prompt-toolkit = "^3.0.36"
jsonschema = "^4.17.3"
ruamel-yaml = "^0.17.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
autopep8 = "^2.0.1"
pre-commit = "^2.21.0"
mypy = "^0.991"
types-pillow = "^9.4.0.0"
types-pyyaml = "^6.0.12.2"
types-pygments = "^2.14.0.0"
types-colorama = "^0.4.15.4"
types-commonmark = "^0.9.2"
coverage = "^7.0.4"
pytest-cov = "^4.0.0"
types-jsonschema = "^4.17.0.3"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"
flake8-pyproject = "^1.2.2"
flake8-simplify = "^0.19.3"

[tool.autopep8]
ignore = "E402"
max_line_length = 100
recursive = true
in-place = true

[tool.flake8]
ignore = "E402, E501, E741, W504, W503, F841"
max-line-length = 100
exclude = "__init__.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
